Add basic rendering
The strategy I settled on is to write primitive functions in
JavaScript and make them available in Rust. I chose this approach
because the arguments and return values for those functions have to go
over the WASM/JavaScript boundary, and the tools I'm using only
support a few values doing that (basically numers, strings, and
structs thereof). As such, I think it will be easier to keep the data
on the Rust side and provide functions to deal with it there while
exposing a few `void` functions to the JavaScript side to handle
things like timing.

I think the only things I'll need to provide from JavaScript are:

- `draw_rect`
- `play_sound`
- User input

I'm confident that I can manage the first two, and I think I can do
the last one by setting an `onclick` callback for the canvas and
handle it in the code that manages the EmbeddedGame.

On the Rust side, I'll have to:

- Do any necessary coordinate conversions (though I think I can keep
  these minimal by choosing the Canvas size appropriately).
- Do updates (mostly deferred to `libsmashbing`, but I'll have to
  handle commands and effects).
- Render the game (using `draw_rect`)
- Trigger sounds (using `play_sound`)
- Handle the `Exit` effect

On the JavaScript side, I'll have to:

- Provide `draw_rect` and `play_sound`.
- Listen for clicks on the Canavas element and indicate them to the EmbeddedGame.

I think I can do all this? And the game is recognizeable in a browser!
I'm cautiously excited. 🤓